separated prior tests, updating config
