+++ test_autofit/mock.py
@@ -3,7 +3,6 @@ import math
 import typing
 
 import autofit as af
-
 # noinspection PyAbstractClass
 import autofit.mapper.prior_model.attribute_pair
 from autofit import Paths
@@ -13,15 +12,15 @@ from autofit.tools.phase import Dataset
 
 class MockNonLinearOptimizer(AbstractOutput):
     def __init__(
-        self,
-        phase_name,
-        phase_tag=None,
-        phase_folders=tuple(),
-        most_probable=None,
-        model_mapper=None,
-        most_likely=None,
-        model_upper_params=None,
-        model_lower_params=None,
+            self,
+            phase_name,
+            phase_tag=None,
+            phase_folders=tuple(),
+            most_probable=None,
+            model_mapper=None,
+            most_likely=None,
+            model_upper_params=None,
+            model_lower_params=None,
     ):
         super(MockNonLinearOptimizer, self).__init__(
             model_mapper or af.ModelMapper(),
@@ -164,11 +163,11 @@ class DeferredClass:
 
 class Galaxy:
     def __init__(
-        self,
-        light_profiles: list = None,
-        mass_profiles: list = None,
-        redshift=None,
-        **kwargs
+            self,
+            light_profiles: list = None,
+            mass_profiles: list = None,
+            redshift=None,
+            **kwargs
     ):
         self.redshift = redshift
         self.light_profiles = light_profiles
@@ -295,13 +294,13 @@ class EllipticalLP(EllipticalProfile):
 
 class AbstractEllipticalSersic(EllipticalProfile):
     def __init__(
-        self,
-        centre=(0.0, 0.0),
-        axis_ratio=1.0,
-        phi=0.0,
-        intensity=0.1,
-        effective_radius=0.6,
-        sersic_index=4.0,
+            self,
+            centre=(0.0, 0.0),
+            axis_ratio=1.0,
+            phi=0.0,
+            intensity=0.1,
+            effective_radius=0.6,
+            sersic_index=4.0,
     ):
         """ Abstract base class for an elliptical Sersic profile, used for computing
         its effective radius and Sersic instance.
@@ -364,13 +363,13 @@ class EllipticalMassProfile(EllipticalProfile, MassProfile):
 # noinspection PyAbstractClass
 class EllipticalCoredPowerLaw(EllipticalMassProfile, MassProfile):
     def __init__(
-        self,
-        centre=(0.0, 0.0),
-        axis_ratio=1.0,
-        phi=0.0,
-        einstein_radius=1.0,
-        slope=2.0,
-        core_radius=0.01,
+            self,
+            centre=(0.0, 0.0),
+            axis_ratio=1.0,
+            phi=0.0,
+            einstein_radius=1.0,
+            slope=2.0,
+            core_radius=0.01,
     ):
         """
         Represents a cored elliptical power-law density distribution
@@ -400,12 +399,12 @@ class EllipticalCoredPowerLaw(EllipticalMassProfile, MassProfile):
 # noinspection PyAbstractClass
 class EllipticalCoredIsothermal(EllipticalCoredPowerLaw):
     def __init__(
-        self,
-        centre=(0.0, 0.0),
-        axis_ratio=1.0,
-        phi=0.0,
-        einstein_radius=1.0,
-        core_radius=0.05,
+            self,
+            centre=(0.0, 0.0),
+            axis_ratio=1.0,
+            phi=0.0,
+            einstein_radius=1.0,
+            core_radius=0.05,
     ):
         """
         Represents a cored elliptical isothermal density distribution, which is
@@ -434,13 +433,13 @@ class EllipticalCoredIsothermal(EllipticalCoredPowerLaw):
 
 class EllipticalSersic(AbstractEllipticalSersic, EllipticalLP):
     def __init__(
-        self,
-        centre=(0.0, 0.0),
-        axis_ratio=1.0,
-        phi=0.0,
-        intensity=0.1,
-        effective_radius=0.6,
-        sersic_index=4.0,
+            self,
+            centre=(0.0, 0.0),
+            axis_ratio=1.0,
+            phi=0.0,
+            intensity=0.1,
+            effective_radius=0.6,
+            sersic_index=4.0,
     ):
         """ The elliptical Sersic profile, used for fitting a model_galaxy's light.
 
@@ -467,17 +466,17 @@ class EllipticalSersic(AbstractEllipticalSersic, EllipticalLP):
 
 class EllipticalCoreSersic(EllipticalSersic):
     def __init__(
-        self,
-        centre=(0.0, 0.0),
-        axis_ratio=1.0,
-        phi=0.0,
-        intensity=0.1,
-        effective_radius=0.6,
-        sersic_index=4.0,
-        radius_break=0.01,
-        intensity_break=0.05,
-        gamma=0.25,
-        alpha=3.0,
+            self,
+            centre=(0.0, 0.0),
+            axis_ratio=1.0,
+            phi=0.0,
+            intensity=0.1,
+            effective_radius=0.6,
+            sersic_index=4.0,
+            radius_break=0.01,
+            intensity_break=0.05,
+            gamma=0.25,
+            alpha=3.0,
     ):
         """ The elliptical cored-Sersic profile, used for fitting a model_galaxy's
         light.
@@ -519,12 +518,12 @@ class EllipticalCoreSersic(EllipticalSersic):
 
 class EllipticalExponential(EllipticalSersic):
     def __init__(
-        self,
-        centre=(0.0, 0.0),
-        axis_ratio=1.0,
-        phi=0.0,
-        intensity=0.1,
-        effective_radius=0.6,
+            self,
+            centre=(0.0, 0.0),
+            axis_ratio=1.0,
+            phi=0.0,
+            intensity=0.1,
+            effective_radius=0.6,
     ):
         """ The elliptical exponential profile, used for fitting a model_galaxy's light.
 
@@ -552,7 +551,7 @@ class EllipticalExponential(EllipticalSersic):
 
 class EllipticalGaussian(EllipticalLP):
     def __init__(
-        self, centre=(0.0, 0.0), axis_ratio=1.0, phi=0.0, intensity=0.1, sigma=0.01
+            self, centre=(0.0, 0.0), axis_ratio=1.0, phi=0.0, intensity=0.1, sigma=0.01
     ):
         """ The elliptical Gaussian profile.
 
@@ -607,3 +606,15 @@ class MockDataset(Dataset):
     @property
     def name(self) -> str:
         return "name"
+
+
+class MockClassGaussian:
+    def __init__(self, one, two):
+        self.one = one
+        self.two = two
+
+
+class MockClassInf:
+    def __init__(self, one, two):
+        self.one = one
+        self.two = two
