+++ test_autofit/test_files/config/json_priors/test_model_mapper.json
@@ -38,46 +38,6 @@
             "upper_limit": 1.0
         }
     },
-    "MockClassGaussian": {
-        "one": {
-            "type": "Gaussian",
-            "lower_limit": 0.0,
-            "upper_limit": 1.0,
-            "width_modifier": {
-                "type": "Absolute",
-                "value": 1.0
-            },
-            "mean": 0.0,
-            "sigma": 1.0
-        },
-        "two": {
-            "type": "Gaussian",
-            "lower_limit": 0.0,
-            "upper_limit": 2.0,
-            "width_modifier": {
-                "type": "Absolute",
-                "value": 2.0
-            },
-            "mean": 0.0,
-            "sigma": 1.0
-        }
-    },
-    "MockClassInf": {
-        "one": {
-            "type": "Gaussian",
-            "lower_limit": "-inf",
-            "upper_limit": 0.0,
-            "mean": 0.0,
-            "sigma": 1.0
-        },
-        "two": {
-            "type": "Gaussian",
-            "lower_limit": 0.0,
-            "upper_limit": "inf",
-            "mean": 0.0,
-            "sigma": 1.0
-        }
-    },
     "ExtendedMockClass": {
         "three": {
             "type": "Uniform",
