--- test_autofit/test_files/config/priors/default/geometry_profiles.ini
@@ -1,18 +1,0 @@
-[GeometryProfile]
-centre_0 = u,0.0,1.0
-centre_1 = u,0.0,1.0
-
-[SphericallProfile]
-
-[EllipticalProfile]
-axis_ratio = u,0.0,2.0
-phi = u,0.0,2.0
-
-[AbstractEllipticalSersic]
-centre_0 = u,-1.0,1.0
-centre_1 = u,-1.0,1.0
-axis_ratio = u,0.0,1.0
-phi = u,0.0,180.0
-intensity = u,0.0,1.0
-effective_radius = u,0.0,3.0
-sersic_index = u,0.6,8.0
