--- test_autofit/test_files/config/priors/default/mock.ini
@@ -1,120 +1,0 @@
-[Redshift]
-redshift = u,0.0,3.0
-
-[RelativeWidth]
-one = u,0,1
-two = u,0,1
-three = u,0,1
-
-[HyperGalaxy]
-contribution_factor = u,1,2
-noise_factor = u,1,2
-noise_power = u,1,2
-
-[EllipticalSersic]
-intensity = u,0.0,1.0
-
-[AbstractEllipticalSersic]
-centre_0 = u,0.0,1.0
-centre_1 = u,0.0,0.5
-axis_ratio = u,0.0,0.5
-phi = u,0.0,0.5
-intensity = g,0.0,0.5
-effective_radius = u,0.0,2.0
-sersic_index = u,0.0,2.0
-mass_to_light_ratio = u,0.0,10.0
-
-[EllipticalExponential]
-centre_0 = u,0.0,1.0
-centre_1 = u,0.0,0.5
-axis_ratio = u,0.0,0.5
-phi = u,0.0,0.5
-intensity = g,0.0,0.5
-effective_radius = u,0.0,2.0
-mass_to_light_ratio = u,0.0,10.0
-
-[SphericalIsothermal]
-centre_0 = u,0.0,1.0
-centre_1 = u,0.0,0.5
-einstein_radius = u,0.0,2.0
-
-[SphericalNFW]
-centre_0 = u,0.0,1.0
-centre_1 = u,0.0,0.5
-kappa_s = u,0.0,2.0
-scale_radius = u,0.0,2.0
-
-[EllipticalGeneralizedNFW]
-kappa_s = u,0.0,1.0
-inner_slope = u,0.0,1.0
-scale_radius = u,0.0,1.0
-
-[EllipticalCoredIsothermal]
-einstein_radius = u,0.0,1.0
-core_radius = u,0.0,1.0
-
-[EllipticalCoreSersic]
-radius_break = u,0.0,1.0
-intensity_break = u,0.0,1.0
-gamma = u,0.0,1.0
-alpha = u,0.0,1.0
-
-[GeometryProfile]
-centre_0 = u,0.0,1.0
-centre_1 = u,0.0,1.0
-
-[SphericallProfile]
-
-[EllipticalProfile]
-axis_ratio = u,0.0,2.0
-phi = u,0.0,2.0
-
-[SimpleClass]
-one = u,0.0,1.0
-two = u,0.0,1.0
-
-
-[DistanceClass]
-first = u,0.0,1.0
-second = u,0.0,1.0
-
-
-[PositionClass]
-position_0 = u,0.0,1.0
-position_1 = u,0.0,1.0
-
-
-[MockClassNLOx4]
-one = u,-120,120
-two = u,-120,120
-three = u,-120,120
-four = u,-120,120
-
-[MockClassNLOx5]
-one = u,-100,101
-two = u,-100,101
-three = u,-100,101
-four = u,-100,101
-five = u,-100,101
-
-[MockClassNLOx6]
-one_0 = u,-100,101
-one_1 = u,-100,101
-two_0 = u,-100,101
-two_1 = u,-100,101
-three = u,-100,101
-four = u,-100,101
-
-[Galaxy]
-redshift = u,0,1
-
-[DeferredClass]
-one = u,0,1
-two = d
-
-[Tracer]
-grid = d
-
-[Circle]
-circumference = u, 0, 2
-radius = u, 0, 2
