--- test_autofit/test_files/config/priors/default/test_model_mapper.ini
@@ -1,30 +1,0 @@
-[MockClassMM]
-one = u,0,1
-two = u,0,1
-
-[MockClassMMinstance]
-one = c,3
-two = u,0,1
-
-[MockClassGaussian]
-one = g,0,1
-two = g,0,1
-
-[MockClassInf]
-one = g,0,1
-two = g,0,1
-
-[ExtendedMockClass]
-three = u,0,1
-
-[MockProfile]
-centre_0 = u,0,1
-centre_1 = u,0,1
-intensity = u,0,1
-
-[WithFloat]
-value = u,0,1
-
-[WithTuple]
-tup_0 = u,0,1
-tup_1 = u,0,1
