--- test_autofit/test_files/config/priors/limit/mock.ini
@@ -1,42 +1,0 @@
-[EllipticalExponential]
-intensity = 0.0,10.0
-
-[EllipticalSersic]
-intensity = 0.0,10.0
-
-[RelativeWidth]
-one = -inf,inf
-two = -inf,inf
-three = -inf,inf
-
-[MockClassNLOx4]
-one = -120,120
-two = -120,120
-three = -120,120
-four = -120,120
-
-[MockClassNLOx5]
-one = -100,101
-two = -100,101
-three = -100,101
-four = -100,101
-five = -100,101
-
-[MockClassNLOx6]
-one_0 = -100,101
-one_1 = -100,101
-two_0 = -100,101
-two_1 = -100,101
-three = -100,101
-four = -100,101
-
-[GeometryProfile]
-centre_0 = -inf,inf
-centre_1 = -inf,inf
-
-[EllipticalProfile]
-axis_ratio = 0.0,1.0
-phi = 0.0,180.0
-
-[Galaxy]
-redshift = -inf,inf
