--- test_autofit/test_files/config/priors/limit/test_model_mapper.ini
@@ -1,16 +1,0 @@
-[MockClassGaussian]
-one = 0,1
-two = 0,2
-
-[MockClassInf]
-one = -inf,0
-two = 0,inf
-
-[MockClassMM]
-one = -10, 10
-two = -10, 10
-
-[MockProfile]
-centre_0 = 0,1
-centre_1 = 0,1
-intensity = 0,1
