--- test_autofit/test_files/config/priors/width/mock.ini
@@ -1,40 +1,0 @@
-[EllipticalExponential]
-intensity = a,1
-
-[RelativeWidth]
-one = r,0.1
-two = r,0.5
-three = r,1.0
-
-
-[MockClassNLOx4]
-one = a,1
-two = a,2
-three = a,2
-four = a,2
-
-[MockClassNLOx5]
-one = a,1
-two = a,2
-three = a,2
-four = a,2
-five = a,2
-
-[MockClassNLOx6]
-one_0 = a,1
-one_1 = a,1
-two_0 = a,2
-two_1 = a,2
-three = a,3
-four = a,4
-
-[GeometryProfile]
-centre_0 = a,0.05
-centre_1 = a,0.05
-
-[EllipticalProfile]
-axis_ratio = a,0.1
-phi = a,10.0
-
-[Galaxy]
-redshift = a,2
