--- test_autofit/test_files/config/priors/width/test_model_mapper.ini
@@ -1,12 +1,0 @@
-[MockClassMM]
-one = a,1
-two = a,2
-
-[MockProfile]
-centre_0 = a,0
-centre_1 = a,0
-intensity = a,0
-
-[MockClassGaussian]
-one = a,1
-two = a,2
